fixed a bug in PatientAppointments, fixed a conflict in business logic (sickness description in PattientAppointment table became the dr's specialization

+++ app/patient_appointment_handler.py
@@ -21,12 +21,12 @@ class PatientAppointmentHandler:
         patient_appointments = db.session.query(PatientAppointment).filter(PatientAppointment.patient_id == patient_id).all()
         return patient_appointments
 
-    def create_normal_appointment(self, patient_id, appointtment_id, sickness):
+    def create_normal_appointment(self, patient_id, appointment_id, specialization):
         """
         brief        : creates a normal appointment 
         param        : patient_id -- int -- unique id of the patient
                        appointment_id -- int -- unique id for the selected appointment 
-                       sickness -- string -- cause for this appointment
+                       specialization -- string -- cardiologist, dentist, ...etc
                        appointment_type -- string -- 'free' or 'reserved'
         constraint   : none
         throws       : none
@@ -35,27 +35,30 @@ class PatientAppointmentHandler:
         """
 
         appointment_handler = AppointmentHandler()
-        appointment = appointment_handler.get_appointment_by_id(appointtment_id)
+        appointment = appointment_handler.get_appointment_by_id(appointment_id)
 
         if appointment is None or appointment.status == "reserved":
             return False
 
         patient_appointment = PatientAppointment(patient_id = patient_id,
-                                                 appointment_id = appointtment_id,
-                                                 sickness = sickness,
+                                                 appointment_id = appointment_id,
+                                                 specialization = specialization,
                                                  appointment_type = "normal")
         
+        db.session.add(patient_appointment)
+        db.session.commit()
+
         # let the appointment status be reserved
         appointment_handler.change_appointment_status(appointment_id)
 
         return True
 
-    def create_urgent_appointment(self, patient_id, sickness):
+    def create_urgent_appointment(self, patient_id, specialization):
         """
         brief        : books an urgent appointment, nearest appointment available today (time wise)
         param        : patient_id -- int -- unique id of the patient 
                        appointment_id -- int -- unique id for the selected appointment 
-                       sickness -- string -- cause for this appointment
+                       specialization -- string -- dentist, cardiologist, ...etc
                        appointment_type -- string -- 'free' or 'reserved'
         constraint   : none
         throws       : none
@@ -71,7 +74,7 @@ class PatientAppointmentHandler:
 
         patient_appointment = PatientAppointment(patient_id = patient_id,
                                                  appointment_id = nearest_appointment_today.appointment_id,
-                                                 sickness = sickness,
+                                                 specialization = specialization,
                                                  appointment_type = "urgent")
 
         db.session.add(patient_appointment)
@@ -96,11 +99,17 @@ class PatientAppointmentHandler:
         appointment_handler = AppointmentHandler()
         appointment = appointment_handler.get_appointment_by_id(appointment_id)
 
+        # check if appointment exists in the system
         if appointment is None:
             return False
 
         patient_appointment = db.session.query(PatientAppointment).filter(PatientAppointment.patient_id == patient_id, 
                                                                           PatientAppointment.appointment_id == appointment_id).first()
+        
+        # check if patient has such an appointment
+        if patient_appointment is None:
+            return False
+
         db.session.delete(patient_appointment)
         db.session.commit()
 
@@ -129,8 +138,11 @@ class PatientAppointmentHandler:
 
         old_pateint_appointment = db.session.query(PatientAppointment).filter(PatientAppointment.appointment_id == old_appointment_id).first()
         
+        if old_pateint_appointment == None:
+            return False
+
         # create the new appointment (with the new id)
-        created = self.create_normal_appointment(old_pateint_appointment.patient_id, new_appointment_id, old_pateint_appointment.sickness)
+        self.create_normal_appointment(old_pateint_appointment.patient_id, new_appointment_id, old_pateint_appointment.specialization)
         
         # delete the old appointment
         self.delete_appointment(old_pateint_appointment.patient_id, old_appointment_id)
@@ -147,21 +159,22 @@ print('---p appoints for p_id=1---')
 for p_appoint in p_appoints:
     print(p_appoint)
 
+
 booked = handler.create_normal_appointment(1, 3, "cold")
-print('appointment (od id=3) booked? ', booked)
+print('appointment (of id=3) booked? ', booked)
 p_appoints = handler.get_all_patient_appointments(1)
 print('---p appoints for p_id=1---')
 for p_appoint in p_appoints:
     print(p_appoint)
 
 deleted = handler.delete_appointment(1, 3)
-print('appointment (od id=3) deleted? ', deleted)
+print('appointment (of id=3) deleted? ', deleted)
 print('---p appoints for p_id=1---')
 p_appoints = handler.get_all_patient_appointments(1)
 for p_appoint in p_appoints:
     print(p_appoint)
 
-booked = handler.create_normal_appointment(1, 4, "cold")
+booked = handler.create_normal_appointment(1, 4, "dentist")
 print('appointment (of id=4) booked? ', booked)
 p_appoints = handler.get_all_patient_appointments(1)
 print('---p appoints for p_id=1---')
@@ -169,7 +182,7 @@ for p_appoint in p_appoints:
     print(p_appoint)
 
 updated  = handler.update_appointment(4, 3)
-print('appointment updated from id=3 to id=4? ', updated)
+print('appointment updated from id=4 to id=3? ', updated)
 p_appoints = handler.get_all_patient_appointments(1)
 print('---p appoints for p_id=1---')
 for p_appoint in p_appoints:
@@ -185,7 +198,7 @@ appoint = Appointment(dr_id=1, start_date=now, end_date=now, status="free")
 db.session.add(appoint)
 db.session.commit()
 
-booked = handler.create_urgent_appointment(1, "cough")
+booked = handler.create_urgent_appointment(1, "cardiologist")
 print('appointment urgent appointment booked? ', booked)
 p_appoints = handler.get_all_patient_appointments(1)
 print('---p appoints for p_id=1---')
